requirements.txt
# KuruSwap Telegram Bot Requirements

# Telegram Bot API
python-telegram-bot==20.7

# Web3 and Ethereum interaction
web3==6.11.3
eth-account==0.9.0

# HTTP requests
requests==2.31.0

# Async support
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0